Renamed sass modules
