+++ src/components/common/Layout/AuthLayout/AuthLayout.tsx
@@ -1,13 +1,13 @@
 import { FC, ReactNode } from 'react'
 
-import styles from './styles.module.sass'
+import styles from './AuthLayout.module.sass'
 
 interface IAuthLayoutProps {
-	children: ReactNode
+  children: ReactNode
 }
 
 const AuthLayout: FC<IAuthLayoutProps> = ({ children }: IAuthLayoutProps) => {
-	return <section className={styles.auth_layout}>{children}</section>
+  return <section className={styles.auth_layout}>{children}</section>
 }
 
 export default AuthLayout
