+++ src/components/common/UI/Form/Button/SubmitBtn/SubmitBtn.tsx
@@ -1,17 +1,17 @@
 import { FC } from 'react'
 
-import styles from './styles.module.sass'
+import styles from './SubmitBtn.module.sass'
 
 interface ISubmitBtnProps {
-	text: string
+  text: string
 }
 
 const SubmitBtn: FC<ISubmitBtnProps> = ({ text }) => {
-	return (
-		<button className={styles.submit} type='submit'>
-			{text}
-		</button>
-	)
+  return (
+    <button className={styles.submit} type='submit'>
+      {text}
+    </button>
+  )
 }
 
 export default SubmitBtn
