+++ src/components/common/UI/Form/FormGroup/FormGroup.module.sass
@@ -1,4 +1,4 @@
-.group
+.form_group
 	display: flex
 	flex-direction: column
