+++ src/components/common/UI/Form/FormGroup/FormGroup.tsx
@@ -1,9 +1,9 @@
 import { FC } from 'react'
 
-import styles from './styles.module.sass'
+import styles from './FormGroup.module.sass'
 
 const FormGroup: FC = ({ children }) => {
-	return <div className={styles.group}>{children}</div>
+  return <div className={styles.form_group}>{children}</div>
 }
 
 export default FormGroup
