+++ src/components/common/UI/Form/FormHeader/FormHeader.module.sass
@@ -1,7 +1,10 @@
-.top
+.header
 	margin-bottom: 50px
 
 	text-align: center
 
 	.title
 		font-size: 36px
+
+	.tip
+		margin-top: 10px
