+++ src/components/common/UI/Form/Label/Label.tsx
@@ -1,18 +1,18 @@
 import { FC } from 'react'
 
-import styles from './styles.module.sass'
+import styles from './Label.module.sass'
 
 interface ILabelProps {
-	for: string
-	text: string
+  for: string
+  text: string
 }
 
 const Label: FC<ILabelProps> = (props: ILabelProps) => {
-	return (
-		<label className={styles.label} htmlFor={props.for}>
-			{props.text}
-		</label>
-	)
+  return (
+    <label className={styles.label} htmlFor={props.for}>
+      {props.text}
+    </label>
+  )
 }
 
 export default Label
