+++ src/components/common/UI/Form/Link/AuthLink/AuthLink.tsx
@@ -1,19 +1,19 @@
 import { FC } from 'react'
 import { Link } from 'react-router-dom'
 
-import styles from './styles.module.sass'
+import styles from './AuthLink.module.sass'
 
 interface ILinkProps {
-	title: string
-	path: string
+  title: string
+  path: string
 }
 
 const AuthLink: FC<ILinkProps> = (props: ILinkProps) => {
-	return (
-		<Link className={styles.link} to={props.path}>
-			{props.title}
-		</Link>
-	)
+  return (
+    <Link className={styles.link} to={props.path}>
+      {props.title}
+    </Link>
+  )
 }
 
 export default AuthLink
