+++ src/components/common/UI/Form/Link/ForgotPassLink/ForgotPassLink.tsx
@@ -1,14 +1,14 @@
 import { FC } from 'react'
 import { Link } from 'react-router-dom'
 
-import styles from './styles.module.sass'
+import styles from './ForgotLink.module.sass'
 
 const ForgotPassLink: FC = () => {
-	return (
-		<Link className={styles.link} to='/password_recovery'>
-			Забыли пароль?
-		</Link>
-	)
+  return (
+    <Link className={styles.link} to='/password_recovery'>
+      Забыли пароль?
+    </Link>
+  )
 }
 
 export default ForgotPassLink
