+++ src/components/pages/Auth/Password/PasswordConfirm/PasswordConfirm.tsx
@@ -10,7 +10,7 @@ import AuthLayout from '../../../../common/Layout/AuthLayout/AuthLayout'
 import Container from '../../../../common/Container/Container'
 import { CustomForm as Form } from '../../../../common/UI/Form'
 
-import styles from '../../styles.module.sass'
+import styles from '../../Auth.module.sass'
 
 const PasswordConfirm: FC = () => {
   const { search } = useLocation()
